update base client and add services

+++ src/main/java/com/baidubce/services/ros/model/FastOrderServiceJob.java
@@ -1,5 +1,14 @@
 /*
- * Copyright (C) 2019 Baidu, Inc. All Rights Reserved.
+ * Copyright 2019 Baidu, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
  */
 package com.baidubce.services.ros.model;
 
@@ -21,7 +30,7 @@ import lombok.NoArgsConstructor;
 public class FastOrderServiceJob {
     private String serviceJobId;
     private LocationPoint location;
-    private Integer serviceStayDuration;
+    private Double serviceStayDuration;
     private Capacity demand;
     private List<TimeWindow> serviceTimeWindows;
 }
